@@ -1,8 +1,8 @@
 library jdenticon_dart;
 
-import 'package:crypto/crypto.dart' show sha1;
 import 'dart:convert' show utf8;
 
+import 'package:crypto/crypto.dart' show sha1;
 import 'src/color.dart';
 import 'src/config.dart';
 import 'src/icon_generator.dart';
@@ -22,21 +22,23 @@ class Jdenticon {
   /// The [padding] is also optional; the default is no padding and it is generally not needed as SvgPicture or the parent widget controlling it, will set the paddings and margins.
   /// To make customized icons, override the values of the settings you want to change: [colorLightnessMinValue], [colorLightnessMaxValue], [grayscaleLightnessMinValue], [grayscaleLightnessMaxValue], [colorSaturation], [grayscaleSaturation], [backColor], [hues]
   /// Note that a given [backColor] should be in #rrggbbaa format and [hues] should be an array with one (or more) int values
-  static String toSvg(String message,
-      {int size = 64,
-      double padding,
-      double colorLightnessMinValue = 0.4,
-      double colorLightnessMaxValue = 0.8,
-      double grayscaleLightnessMinValue = 0.3,
-      double grayscaleLightnessMaxValue = 0.9,
-      double colorSaturation = 0.5,
-      double grayscaleSaturation = 0.0,
-      String backColor = '#FFFFFF',
-      List<int> hues}) {
+  static String toSvg(
+    String message, {
+    int size = 64,
+    double padding = 0.08,
+    double colorLightnessMinValue = 0.4,
+    double colorLightnessMaxValue = 0.8,
+    double grayscaleLightnessMinValue = 0.3,
+    double grayscaleLightnessMaxValue = 0.9,
+    double colorSaturation = 0.5,
+    double grayscaleSaturation = 0.0,
+    String backColor = '#FFFFFF',
+    List<int> hues = const <int>[],
+  }) {
     final String hash = '${sha1.convert(utf8.encode(message))}';
-    SvgWriter writer = SvgWriter(size.abs());
+    final SvgWriter writer = SvgWriter(size.abs());
     final double s = size.abs().toDouble();
-    SvgRenderer renderer = SvgRenderer(writer);
+    final SvgRenderer renderer = SvgRenderer(writer);
     IconGenerator(
         renderer,
         hash,
@@ -57,21 +59,22 @@ class Jdenticon {
   }
 
   /// Returns the current configuration constant settings used by Jdenticon
-  static Config getCurrentConfig(
-      {double colorLightnessMinValue = 0.4,
-      double colorLightnessMaxValue = 0.8,
-      double grayscaleLightnessMinValue = 0.3,
-      double grayscaleLightnessMaxValue = 0.9,
-      double colorSaturation = 0.5,
-      double grayscaleSaturation = 0.0,
-      String backColor = '#FFFFFF',
-      List<int> hues}) {
-    Function lightness(
+  static Config getCurrentConfig({
+    double colorLightnessMinValue = 0.4,
+    double colorLightnessMaxValue = 0.8,
+    double grayscaleLightnessMinValue = 0.3,
+    double grayscaleLightnessMaxValue = 0.9,
+    double colorSaturation = 0.5,
+    double grayscaleSaturation = 0.0,
+    String backColor = '#FFFFFF',
+    List<int> hues = const <int>[],
+  }) {
+    double Function(double) lightness(
         String configName, double defaultMin, double defaultMax) {
-      List<double> range = [defaultMin, defaultMax];
+      final List<double> range = [defaultMin, defaultMax];
 
       return (double value) {
-        double value2 = range[0] + value * (range[1] - range[0]);
+        final double value2 = range[0] + value * (range[1] - range[0]);
         return value2 < 0.0 ? 0.0 : (value2 > 1.0 ? 1.0 : value2);
       };
     }
