@@ -1,10 +1,10 @@
 import 'dart:math' show Point;
 
 abstract class Renderer {
-  setBackground(String fillColor);
-  beginShape(String color);
-  endShape();
-  addPolygon(List<Point> points);
-  addCircle(Point point, double diameter, bool counterClockwise);
-  finish();
+  void setBackground(String fillColor);
+  void beginShape(String color);
+  void endShape();
+  void addPolygon(List<Point> points);
+  void addCircle(Point point, double diameter, {bool counterClockwise});
+  void finish();
 }
