@@ -1,8 +1,6 @@
 import 'dart:math' show Point;
 
-double svgValue(num value) {
-  return ((value * 10 + 0.5).floor() / 10);
-}
+double svgValue(num value) => (value * 10 + 0.5).floor() / 10;
 
 class SvgPath {
   String dataString = '';
@@ -10,21 +8,23 @@ class SvgPath {
   SvgPath();
 
   void addPolygon(List<Point> points) {
-    String ds =
-        'M${svgValue(points[0].x).toStringAsFixed(1)} ${svgValue(points[0].y).toStringAsFixed(1)}';
+    final StringBuffer ds = StringBuffer(
+        'M${svgValue(points[0].x).toStringAsFixed(1)} ${svgValue(points[0].y).toStringAsFixed(1)}');
     for (int i = 0; i < points.length; i++) {
-      ds +=
-          "L${svgValue(points[i].x).toStringAsFixed(1)} ${svgValue(points[i].y).toStringAsFixed(1)}";
+      ds.write(
+          "L${svgValue(points[i].x).toStringAsFixed(1)} ${svgValue(points[i].y).toStringAsFixed(1)}");
     }
-    this.dataString += ds + "Z";
+    ds.write("Z");
+    dataString += ds.toString();
   }
 
-  void addCircle(Point point, double diameter, bool counterClockwise) {
-    int sweepFlag = counterClockwise != null ? (counterClockwise ? 0 : 1) : 1;
-    int svgRadius = (svgValue(diameter / 2)).floor();
-    int svgDiameter = (svgValue(diameter)).floor();
+  void addCircle(Point point, double diameter,
+      {bool counterClockwise = false}) {
+    final int sweepFlag = counterClockwise ? 0 : 1;
+    final int svgRadius = (svgValue(diameter / 2)).floor();
+    final int svgDiameter = (svgValue(diameter)).floor();
 
-    this.dataString +=
+    dataString +=
         'M${svgValue(point.x)} ${svgValue(point.y + diameter / 2)}a$svgRadius,$svgRadius 0 1,$sweepFlag $svgDiameter ,0a$svgRadius,$svgRadius 0 1,$sweepFlag -$svgDiameter,0';
   }
 }
