@@ -1,28 +1,30 @@
 import 'dart:math' show Point;
 
 class Transform {
-  double _x;
-  double _y;
-  double _size;
-  double _rotation;
+  final double _x;
+  final double _y;
+  final double _size;
+  final double _rotation;
 
   Transform(this._x, this._y, this._size, this._rotation);
 
-  Point transformPoint(double x, double y, [double w, double h]) {
-    double right = this._x + this._size;
-    double bottom = this._y + this._size;
-    double height = h != null ? h : 0;
-    double width = w != null ? w : 0;
-    return this._rotation == 1
-        ? Point(right - y - height, this._y + x)
-        : this._rotation == 2
+  Point transformPoint(double x, double y,
+      [double width = 0.0, double height = 0.0]) {
+    final double right = _x + _size;
+    final double bottom = _y + _size;
+
+    return _rotation == 1
+        ? Point(right - y - height, _y + x)
+        : _rotation == 2
             ? Point(right - x - width, bottom - y - height)
-            : this._rotation == 3
-                ? Point(this._x + y, bottom - x - width)
-                : Point(this._x + x, this._y + y);
+            : _rotation == 3
+                ? Point(_x + y, bottom - x - width)
+                : Point(_x + x, _y + y);
   }
 
-  static Transform noTransform() {
-    return Transform(0, 0, 0, 0);
-  }
+  Transform.noTransform()
+      : _x = 0,
+        _y = 0,
+        _size = 0,
+        _rotation = 0;
 }
